moved server handlers
